backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic>=0.40.0
music21==9.3.0
Pillow>=10.0.0
pdf2image==1.17.0
python-multipart==0.0.12
aiofiles>=24.0.0
httpx>=0.27.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
